+++ api/myapp/__init__.py
@@ -1,2 +1,2 @@
-from .app import app
+# from app import app
