+++ api/myapp/views/db.py
@@ -1,6 +1,6 @@
 # import asyncio
 # import aiomysql
-from myapp import app
+from . import app
 
 async def get_db_connection():
     return await app.get_db()
