+++ api/myapp/views/dealsList_view.py
@@ -1,7 +1,7 @@
 from flask import Blueprint
 import pandas as pd
 
-from ..app import mysql
+from ...app import mysql
 
 
 deals_list = Blueprint("deals_list",__name__)
