+++ api/myapp/views/hello.py
@@ -1,6 +1,6 @@
 from flask import Blueprint,render_template
 
-from ..app import mysql
+from ...app import mysql
 # from ..app import app
 
 hello = Blueprint('hello',__name__)
